class-32/demo/blogpostgres/pyproject.toml
[tool.poetry]
name = "blogpostgres"
version = "0.1.0"
description = ""
authors = ["yahyaalbashar <v.salvatore7.gs@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
